feat: allow adding multiple maven dependencies
