+++ plugins/maven-backend/src/scaffolder/actions/maven/dependenciesAdd.example.ts
@@ -3,19 +3,21 @@ import yaml from 'yaml';
 
 export const examples: TemplateExample[] = [
     {
-        description: 'Add a maven dependency to a project',
+        description: 'Add a maven dependencies to a project',
         example: yaml.stringify({
             steps: [
                 {
-                    action: 'maven:dependency:add',
+                    action: 'maven:dependencies:add',
                     id: 'maven-dependency-add',
                     name: 'Add a dependency to a Maven project',
                     input: {
                         values: {
-                            groupId: 'org.mvnpm.at.mvnpm',
-                            artifactId: 'vaadin-webcomponents',
-                            version: '24.3.10',
-                            scope: 'provided',
+                            dependencies: [{
+                              groupId: 'org.mvnpm.at.mvnpm',
+                              artifactId: 'vaadin-webcomponents',
+                              version: '24.3.10',
+                              scope: 'provided',
+                            }],
                         },
                     },
                 },
